fix(clean): handle a graceful shutdown correctly
Fixes TIMOB-27866
